1.3.3dev: Make numeric literals compatible with Python3
Patch by Jun Omae.

Refs #12130.


git-svn-id: http://trac.edgewall.org/intertrac/log:/trunk@16658 af82e41b-90c4-0310-8c96-b1721e28e2e2

+++ trac/ticket/query.py
@@ -321,7 +321,7 @@ class Query(object):
                         if href is not None:
                             result['href'] = href.ticket(val)
                     elif name in self.time_fields:
-                        val = from_utimestamp(long(val)) if val else None
+                        val = from_utimestamp(int(val)) if val else None
                     elif field and field['type'] == 'checkbox':
                         val = as_bool(val)
                     elif val is None:
